+++ srcs/u_test.c
@@ -10,6 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
+#include <stdio.h>
 
 int	function_to_test(unsigned int nb, unsigned int nb2)
 {
@@ -19,20 +20,22 @@ int	function_to_test(unsigned int nb, unsigned int nb2)
 int	unit_test2(void)
 {
 	if (function_to_test(6, 6) == 12)
-		return (1);
-	return (0);
+		return (0);
+	return (1);
 }
 
 int unit_test3(void)
 {
-	if (function_to_test(0, 0) == 0)
-		return (1);
-	return (0);
+	 int *ptr = NULL;
+    *ptr = 42;  // Segmentation fault ici !
+    
+    printf("Cette ligne ne sera jamais exécutée\n");
+    return 0;
 }
 
 int	unit_test1(void)
 {
 	if (function_to_test(2, 5) == 4)
-		return (1);
-	return (0);
+		return (0);
+	return (1);
 }
